FadeOutAudio working

+++ LC_Unity/Assets/Scripts/Engine/MusicAndSounds/FadeOutAudio.cs
@@ -6,7 +6,7 @@ namespace Engine.MusicAndSounds
     public abstract class FadeOutAudio : IRunnable
     {
         public string Name { get; set; }
-        public int TransitionDuration { get; set; }
+        public float TransitionDuration { get; set; }
         public UnityEvent Finished { get; set; }
         public bool IsFinished { get; set; }
